+++ .gitignore
@@ -1,2 +1,3 @@
 .env
 /node_modules
+backend/node_modules
