+++ backend/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "backend",
+  "version": "1.0.0",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "description": "",
+  "dependencies": {
+    "apollo-server-express": "^3.13.0",
+    "express": "^5.1.0",
+    "graphql": "^16.10.0",
+    "pg": "^8.14.1",
+    "bcrypt": "^5.1.1",
+    "dotenv": "^16.5.0",
+    "graphql-yoga": "^5.13.4"
+  }
+}
