+++ backend/src/resolvers/index.js
@@ -1,4 +1,4 @@
-import { db } from '../db.js';
+import { db } from '../config/db.js';
 
 export const resolvers = {
   Query: {
